requirements.txt
asgiref==3.2.5
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Django==2.2.9
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
googletrans==2.4.0
idna==2.9
joblib==0.14.1
mysqlclient==1.4.6
nltk==3.5
oauthlib==3.1.0
pip-autoremove==0.9.1
PySocks==1.7.1
python-dotenv==0.13.0
pytz==2019.3
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tqdm==4.45.0
tweepy==3.8.0
Unidecode==1.1.1
urllib3==1.25.9
